Remove unnecessary whitespace and dashes
The previous version of this file used a tabular format like this

```
|------|------|-----|
| a    | b    | c   |
|------|------|-----|
```

i.e. the spacing looks correct if you output the file using a fixed-width font.
But many of the lines in the file are very long, which makes the lines wrap
when displaying. e.g. if you are trying to edit the file using the github web
interface. The `extra` whitespace is not needed in order for the file to
display correctly when rendered by the github web interface, which is the
primary use for this file.

This change removes almost all the whitespace which is not strictly necessary
(I've left a bit for visual clarity, so `| a | b |` instead of `|a|b|`). i.e.

```
|-|-|-|
| a | b | c |
|-|-|-|
```

This makes it much easier to edit the file via the github ui, and also means
the page loads significantly faster.