Fixes the matching of search results
Currently the search will always return any row where the acronym is a
link to another resource.  This means that the result users actually
want is often buried below the fold.

By fixing the rendering of the row when the acronym is a link, this PR
makes the search work the way users would expect.  Rows where the
acronym is a link and are now only shown after a search if the text of
the link is a valid match.

Fixes #83